parachute youhou

+++ src/main/java/fr/formation/dao/IDAOParachutiste.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import fr.formation.model.Parachutiste;
 
-public interface IDAOParachutiste extends JpaRepository<Parachutiste, Integer> {
-
+public interface IDAOParachutiste extends JpaRepository<Parachutiste, Long> {
+	
 }
